requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database (uncomment if needed)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0
# pymongo>=4.6.0

# Data processing (uncomment if needed)
# pandas>=2.1.0
# numpy>=1.26.0

# Logging
loguru>=0.7.0

